refactor(ui): modularize node properties views
