+++ packages/platform-ui/src/components/nodeProperties/viewRegistry.ts
@@ -1,0 +1,16 @@
+import type { NodeKind } from './types';
+import type { NodePropertiesViewComponent } from './viewTypes';
+
+import ToolNodeConfigView from './views/ToolNodeConfigView';
+import WorkspaceNodeConfigView from './views/WorkspaceNodeConfigView';
+import McpNodeConfigView from './views/McpNodeConfigView';
+import AgentNodeConfigView from './views/AgentNodeConfigView';
+import TriggerNodeConfigView from './views/TriggerNodeConfigView';
+
+export const NODE_VIEW_REGISTRY: Record<NodeKind, NodePropertiesViewComponent> = {
+  Tool: ToolNodeConfigView,
+  Workspace: WorkspaceNodeConfigView,
+  MCP: McpNodeConfigView,
+  Agent: AgentNodeConfigView,
+  Trigger: TriggerNodeConfigView,
+};
