+++ packages/platform-ui/src/components/nodeProperties/viewTypes.ts
@@ -1,0 +1,8 @@
+import type { NodePropertiesSidebarProps } from './types';
+
+export interface NodePropertiesViewProps extends NodePropertiesSidebarProps {
+  secretSuggestions: string[];
+  variableSuggestions: string[];
+}
+
+export type NodePropertiesViewComponent = (props: NodePropertiesViewProps) => JSX.Element | null;
